requirements.txt
Flask==2.3.2
redis==4.6.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
pymysql==1.0.3
